Restructured some directories plus some changes to the DB and Product calls
