+++ product/create.go
@@ -1,0 +1,4 @@
+package product
+
+
+//NewProduct()
