+++ product/delete.go
@@ -1,0 +1,5 @@
+package product
+
+
+//Delete Product
+//Archive Product
